feat(ptcc): add the detection module API, module supply and a fractional setpoint
Added:
- rp_Ptcc*LabM* calls covering the detection module panel: parameters and monitor
  for every register, per-module limits, the list of gains the device defines, and
  setters for detector bias voltage and current compensation, output DC offset,
  preamplifier gain, varactor compensation, 1st stage transimpedance, signal
  coupling and bandwidth
- rp_PtccSetSupply() writing the module supply control mode and both rails, and
  rp_PtccSetPwm() writing the TEC PWM setting
- rp_PtccGetLabMMonitor(), a cached LAB_M sample refreshed by the background
  poller alongside the controller monitor
- read_lab_m_params(), read_lab_m_monitor() and lab_m_gain_values() helpers in the
  SWIG module
- Long options in ptcc_control for every new setting: --labm-monitor,
  --labm-params, --gains, --bias-voltage, --bias-current, --offset, --gain,
  --varactor, --transimpedance, --coupling, --bandwidth, --supply-mode,
  --supply-plus, --supply-minus and --pwm, each printing the current value when
  given without one
- tests/test_api_labm.cpp and a LabMTests class in the SWIG suite

Changed:
- rp_PtccSetSetpoint() takes a float instead of whole Kelvins: the protocol
  encodes the setpoint with three decimals, so a tenth of a Kelvin reaches the
  device
- ptcc_emulator.py keeps the writable parameters instead of answering with static
  frames, so a value written by the driver is what the next read returns, and
  serves the LAB_M containers when it presents a LAB_M module

+++ rp-api/api-ptcc/tests/ptcc_diag.py
@@ -31,7 +31,7 @@ def dump(title, values):
 def main():
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("-d", "--device", default=None)
-    parser.add_argument("-t", "--temperature", type=int, default=230)
+    parser.add_argument("-t", "--temperature", type=float, default=230.0)
     parser.add_argument("--set", action="store_true", help="actually write the setpoint (EEPROM)")
     args = parser.parse_args()
